% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVecOcc.R
\name{as.vectorpoints}
\alias{as.vectorpoints}
\title{Convert data.frames to vector.points objects.}
\usage{
as.vectorpoints(x)
}
\arguments{
\item{x}{A data.frame}
}
\description{
Will create empty columns for any missing columns expected in a vector.points data.frame.
This function is particularly useful for use with packages like dplyr that strip
objects of their classes.
}
\examples{
\dontrun{
library(dplyr)
Brazil_vec <- getVecOcc(country = "Brazil")

# Filter data.frame then readd vector points class so autoplot can be used.
Brazil_vec \%>\%
  filter(sample_method1 == 'larval collection') \%>\%
  as.vectorpoints \%>\%
  autoplot
}
}
