% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.default.R
\name{autoplot.default}
\alias{autoplot.default}
\title{Default autoplot method}
\usage{
\method{autoplot}{default}(object, ...)
}
\arguments{
\item{object}{object to plot}

\item{...}{other arguments}
}
\description{
Default autoplot method
}
