% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableData.R
\name{availableData}
\alias{availableData}
\title{Available data to download from the MAP geoserver.}
\usage{
availableData(sourcedata = NULL, full_results = FALSE,
  country = NULL, ISO = NULL, continent = NULL)
}
\arguments{
\item{sourcedata}{One of 'pr points' or 'vector points'}

\item{full_results}{Should the list be printed to the console?}
}
\value{
\code{availableData} returns a data.frame detailing the administrative units for which shapefiles are stored on the MAP geoserver.
}
\description{
\code{availableData} is a wrapper for isAvailable_pr and isAvailable_vec, listing data (PR survey point location data and vector occurrence locations available to download from the MAP geoserver.
}
\examples{
\donttest{
available_pr_locations <- isAvailable_pr()
available_vector_locations <- isAvailable_vec()
}
}
\seealso{
\code{link{isAvailable_pr}}
\code{\link{isAvailable_vec}}
}
