% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{build_cql_bbox_filter}
\alias{build_cql_bbox_filter}
\title{Builds a CQL filter to be used with getFeatures, that will filter based on the given bounding box.}
\usage{
build_cql_bbox_filter(bbox)
}
\arguments{
\item{bbox}{A matrix representing a bounding box.}
}
\value{
The character string of the CQL filter.
}
\description{
Builds a CQL filter to be used with getFeatures, that will filter based on the given bounding box.
}
\keyword{internal}
