% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{build_cql_filter}
\alias{build_cql_filter}
\title{Builds a cql filter to be used with getFeatures, that will filter based on the given list of values.}
\usage{
build_cql_filter(attribute, values)
}
\arguments{
\item{attribute}{A string representing the attribute to filter by}

\item{values}{A character, or character list, representing one or more values}
}
\value{
The character string of the cql filter.
}
\description{
Builds a cql filter to be used with getFeatures, that will filter based on the given list of values.
}
\keyword{internal}
