% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{build_cql_time_filter}
\alias{build_cql_time_filter}
\title{Builds a cql filter to be used with getFeatures, that will filter based on the time range
provided by start_date and end_date.}
\usage{
build_cql_time_filter(start_date, end_date)
}
\arguments{
\item{start_date}{A Date Object that is the lower bound on the time range (inclusive).}

\item{end_date}{A Date Object that is the upper bound on the time range (exclusive).}
}
\value{
The character string of the cql filter.
}
\description{
Builds a cql filter to be used with getFeatures, that will filter based on the time range
provided by start_date and end_date.
}
\keyword{internal}
