% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{callGetFeaturesWithFilters}
\alias{callGetFeaturesWithFilters}
\title{Calls getFeatures on the given type with the given filters, where they are not NULL.}
\usage{
callGetFeaturesWithFilters(feature_type, cql_filter, ...)
}
\arguments{
\item{feature_type}{Object of WFSFeatureType.}

\item{cql_filter}{A character string that reflects cql filter e.g. "time_start AFTER 2020-00-00T00:00:00Z AND country IN ('Nigeria')"}
}
\value{
The features returned from the getFeatures called on the feature_type with the filter where not NULL.
}
\description{
Calls getFeatures on the given type with the given filters, where they are not NULL.
}
\keyword{internal}
