% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listRaster.R
\name{clean_mosquito_names}
\alias{clean_mosquito_names}
\title{Function to clean up a bunch of mess in the mosquito names.}
\usage{
clean_mosquito_names(available_rasters)
}
\arguments{
\item{available_rasters}{data.frame of available rasters, with a column for title..}
}
\value{
The same data.frame inputted, but now with some of the titles corrected.
}
\description{
Function to clean up a bunch of mess in the mosquito names.
}
\keyword{internal}
