% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRaster.R
\name{collectRastersOfSameResolution}
\alias{collectRastersOfSameResolution}
\title{Will group rasters that have the same resolution into a SpatRasterCollection. Ones that do not have the same
resolution as any other raster are left as a single SpatRaster.}
\usage{
collectRastersOfSameResolution(rst_list)
}
\arguments{
\item{rst_list}{list of SpatRasters.}
}
\value{
either a list of SpatRasters/SpatRasterCollections, or a single SpatRaster or SpatRasterCollection
}
\description{
Will group rasters that have the same resolution into a SpatRasterCollection. Ones that do not have the same
resolution as any other raster are left as a single SpatRaster.
}
\keyword{internal}
