% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{combine_cql_filters}
\alias{combine_cql_filters}
\title{Builds a cql filter from a list of cql sub filters}
\usage{
combine_cql_filters(filter_list)
}
\arguments{
\item{filter_list}{List of characters and NULL objects that represent individual cql filters e.g. list("time_start AFTER 2020-00-00T00:00:00Z", NULL, "country IN ('Nigeria')")}
}
\value{
If all values in filter_list are null, will return NULL, else will return character of combined filters. e.g. "time_start AFTER 2020-00-00T00:00:00Z AND country IN ('Nigeria')"
}
\description{
Builds a cql filter from a list of cql sub filters
}
\keyword{internal}
