% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{convert_to_date_with_trycatch}
\alias{convert_to_date_with_trycatch}
\title{Tries to convert character into a Date object. If this fails, the program will be stopped and an error message
shown to the user}
\usage{
convert_to_date_with_trycatch(input, input_name)
}
\arguments{
\item{input}{A character of length 1, to convert into a Date object.}

\item{input_name}{A character string that reflects the name of the input (to inform the user).}
}
\value{
Will return input as a Date is it can be converted into one. Else will stop the program and issue the user with an error message.
}
\description{
Tries to convert character into a Date object. If this fails, the program will be stopped and an error message
shown to the user
}
\keyword{internal}
