% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRaster.R
\name{download_rst}
\alias{download_rst}
\title{Download rasters from the MAP geoserver to a specifed location. If file already exists it will read it instead.}
\usage{
download_rst(dataset_id, extent, year, file_name, file_path)
}
\arguments{
\item{dataset_id}{ID for dataset on MAP geoserver}

\item{extent}{desired raster extent}

\item{year}{desired year to download}

\item{file_name}{file name (excluding extension) to save raster to}

\item{file_path}{path to save raster to}
}
\value{
SpatRaster
}
\description{
Download rasters from the MAP geoserver to a specifed location. If file already exists it will read it instead.
}
