% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShp.R
\name{downloadShp}
\alias{downloadShp}
\title{Download MAPadmin2013 Administrative Boundary Shapefiles from the MAP geoserver}
\usage{
downloadShp(URL)
}
\arguments{
\item{country}{string containing name of desired country, e.g. \code{ c("Country1", "Country2", ...)} OR \code{ = "ALL"} (use either \code{ISO} OR \code{country})}

\item{ISO}{string containing ISO3 code for desired country, e.g. \code{c("XXX", "YYY", ...)} OR \code{ = "ALL"} (use either \code{ISO} OR \code{country})}

\item{admin_level}{string specifying the administrative level for which shapefile are desired (only "admin1", "admin0", or "both" accepted)}
}
\value{
\code{getShp} returns a list containing separate shapefile dataframes for each administrative level. The following attribute fields are included:

\enumerate{
\item \code{COLUMNNAME} description of contents
\item \code{COLUMNNAME} description of contents
\item \code{COLUMNNAME} description of contents
}
}
\description{
\code{getShp} downloads shapefiles for a specified country (or countries) and returns this as a spatial
}
\examples{
#Download PfPR data & associated shapefiles for Nigeria and Cameroon
\dontrun{NGA_CMR_PR <- getPR(country = c("Nigeria", "Cameroon"), species = "Pf")}
\dontrun{NGA_CMR_shp <- getShp(country = c("Nigeria", "Cameroon"))}

#' #Download PfPR data & associated shapefiles for Chad
\dontrun{Chad_PR <- getPR(ISO = "TCD", species = "both")}
\dontrun{Chad_shp <- getShp(ISO = "TCD")}

#' #Download PfPR data & associated shapefiles defined by bbox for Madagascar
\dontrun{MDG_PR <- getPR(country = "Madagascar", species = "Pv")}


}
\seealso{
\code{autoplot} method for quick mapping of PR point locations (\code{\link{autoplot.pr.points}}).
}
