% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRaster.R
\name{extractLayerValues}
\alias{extractLayerValues}
\title{Returns a data.frame of the extracted raster values with columns for layerName, year, value, long and lat.}
\usage{
extractLayerValues(points_to_query, dataset_id, year)
}
\arguments{
\item{points_to_query}{matrix of points to get raster data for, with first column being the latitude and second being the longitude.}

\item{dataset_id}{character representing the dataset id.}

\item{year}{either a single year, or a list of years, or NA.}
}
\description{
Returns a data.frame of the extracted raster values with columns for layerName, year, value, long and lat.
}
\keyword{internal}
