% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{fetchCountriesGivenDatasetId}
\alias{fetchCountriesGivenDatasetId}
\title{Get the list of available countries for a given dataset_id in the Explorer workspace,}
\usage{
fetchCountriesGivenDatasetId(wfs_client, dataset_id)
}
\arguments{
\item{dataset_id}{The dataset id from which to get the list of available countries from.}
}
\value{
A dataframe with columns for country, country_id and continent_id
}
\description{
Get the list of available countries for a given dataset_id in the Explorer workspace,
}
\keyword{internal}
