% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRaster.R
\name{fetchRaster}
\alias{fetchRaster}
\title{Fetches the raster from geoserver, given the dataset id of the raster, and filter value for bbox and year.}
\usage{
fetchRaster(dataset_id, bbox_filter, year)
}
\arguments{
\item{dataset_id}{character that is the dataset id of a raster.}

\item{bbox_filter}{character that represents a bbox filter, or NULL.}

\item{year}{numeric that is a single year, or NA.}
}
\value{
the SpatRaster that matches the dataset_id with filters applied for bbox_filter and year if not NULL/NA.
}
\description{
Fetches the raster from geoserver, given the dataset id of the raster, and filter value for bbox and year.
}
\keyword{internal}
