% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillDHScoordinates.R
\name{fillDHSCoordinates}
\alias{fillDHSCoordinates}
\title{Add DHS locations to malaria data}
\usage{
fillDHSCoordinates(data, email = NULL, project = NULL, cache_path = NULL,
  config_path = NULL, global = TRUE, verbose_download = FALSE,
  verbose_setup = TRUE, data_frame = NULL, timeout = 30,
  password_prompt = FALSE, prompt = TRUE)
}
\arguments{
\item{data}{Data to add DHS coordinates to}
}
\description{
We cannot directly share DHS data. WE can share coordinates,
  but not the data values. This function attempted to fill
  the data gaps directly from the DHS server using the package
  rdhs. To use the function you will need to setup an 
  account on the DHS website and request any datasets you wish
  to use (including requesting the GPS data). Once this has 
  been verified, you should be able to use this function.
}
\examples{
\dontrun{
pf <- malariaAtlas::getPR("all",species = "pf")
pf <- fillDHSCoordinates(pf, 
email = "pretend_email@gmail.com",
project = "pretend project name")
}
}
\author{
OJ Watson
}
