% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{get_name_from_wfs_feature_type_id}
\alias{get_name_from_wfs_feature_type_id}
\title{Get the name from a wfs feature type id.}
\usage{
get_name_from_wfs_feature_type_id(id)
}
\arguments{
\item{id}{The ID to parse.}
}
\value{
A name of the dataset
}
\description{
Get the name from a wfs feature type id.
}
\keyword{internal}
