% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{get_wcs_clients}
\alias{get_wcs_clients}
\title{WCS clients lazily created or from cache}
\usage{
get_wcs_clients(logger = NULL)
}
\value{
A list with of WCSClients by workspace
}
\description{
WCS clients lazily created or from cache
}
\keyword{internal}
