% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{get_wcs_coverage_summary_from_raster_id}
\alias{get_wcs_coverage_summary_from_raster_id}
\title{Get the WCS coverage summary for a raster ID.}
\usage{
get_wcs_coverage_summary_from_raster_id(raster)
}
\arguments{
\item{raster}{The raster ID.}
}
\value{
A \link[ows4R]{WCSCoverageSummary} object.
}
\description{
Get the WCS coverage summary for a raster ID.
}
\keyword{internal}
