% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{get_wfs_clients}
\alias{get_wfs_clients}
\title{WFS clients lazily created or from cache}
\usage{
get_wfs_clients(logger = NULL)
}
\value{
A list with of WFSClients by workspace
}
\description{
WFS clients lazily created or from cache
}
\keyword{internal}
