% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{get_wms_clients}
\alias{get_wms_clients}
\title{WMS clients lazily created or from cache}
\usage{
get_wms_clients(logger = NULL)
}
\value{
A list with of WMSClients by workspace
}
\description{
WMS clients lazily created or from cache
}
\keyword{internal}
