% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{get_workspace_and_version_from_coverage_id}
\alias{get_workspace_and_version_from_coverage_id}
\title{Get the workspace and version from a raster id.}
\usage{
get_workspace_and_version_from_coverage_id(id)
}
\arguments{
\item{id}{The ID to parse.}
}
\value{
A list with the workspace and version.
}
\description{
Get the workspace and version from a raster id.
}
\keyword{internal}
