% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{get_workspace_and_version_from_wfs_feature_type_id}
\alias{get_workspace_and_version_from_wfs_feature_type_id}
\title{Get the workspace and version from a wfs feature type id.}
\usage{
get_workspace_and_version_from_wfs_feature_type_id(id)
}
\arguments{
\item{id}{The ID to parse.}
}
\value{
A list with the workspace and version.
}
\description{
Get the workspace and version from a wfs feature type id.
}
\keyword{internal}
