% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{getLatestDatasetId}
\alias{getLatestDatasetId}
\title{Get the dataset id of the latest version of a given dataset name within a dataset}
\usage{
getLatestDatasetId(datasets, dataset_name)
}
\arguments{
\item{datasets}{The datasets in a data.frame with columns for dataset_id, version and workspace.}

\item{dataset_name}{The dataset name e.g. 'Global_Pf_Parasite_Rate_Surveys'.}
}
\value{
A dataset id within the datasets that matches the dataset name and has the most recent version
}
\description{
Get the dataset id of the latest version of a given dataset name within a dataset
}
\keyword{internal}
