% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{getLatestDatasetIdForPfPrData}
\alias{getLatestDatasetIdForPfPrData}
\title{Get the dataset id of the latest version of pf PR data}
\usage{
getLatestDatasetIdForPfPrData()
}
\value{
The dataset id of the latest version of pf PR Data
}
\description{
Get the dataset id of the latest version of pf PR data
}
\keyword{internal}
