% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{getLatestDatasetIdForPvPrData}
\alias{getLatestDatasetIdForPvPrData}
\title{Get the dataset id of the latest version of pv PR data}
\usage{
getLatestDatasetIdForPvPrData()
}
\value{
The dataset id of the latest version of pv PR Data
}
\description{
Get the dataset id of the latest version of pv PR data
}
\keyword{internal}
