% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{getLatestDatasetIdForVecOccData}
\alias{getLatestDatasetIdForVecOccData}
\title{Get the dataset id of the latest version of vector occurrence data}
\usage{
getLatestDatasetIdForVecOccData()
}
\value{
The dataset id of the latest version of vector occurrence Data
}
\description{
Get the dataset id of the latest version of vector occurrence data
}
\keyword{internal}
