% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{getLatestVersionForAdminData}
\alias{getLatestVersionForAdminData}
\title{Get the latest version of admin boundary data}
\usage{
getLatestVersionForAdminData()
}
\value{
The latest version of admin boundary data
}
\description{
Get the latest version of admin boundary data
}
\keyword{internal}
