% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRaster.R
\name{getLatitudeColumn}
\alias{getLatitudeColumn}
\title{Returns the best latitude column name in df data.frame, if one exists.}
\usage{
getLatitudeColumn(df)
}
\arguments{
\item{df}{The data.frame, hopefully containing a column for latitude}
}
\value{
Returns the best column name for latitude - 'latitude', 'lat', or 'y' respectively, or NULL if none of these column names exists in df
}
\description{
Returns the best latitude column name in df data.frame, if one exists.
}
\keyword{internal}
