% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRaster.R
\name{getLongitudeColumn}
\alias{getLongitudeColumn}
\title{Returns the best longitude column name in df data.frame, if one exists.}
\usage{
getLongitudeColumn(df)
}
\arguments{
\item{df}{The data.frame, hopefully containing a column for longitude.}
}
\value{
Returns the best column name for longitude - 'longitude', 'long', or 'x' respectively, or NULL if none of these column names exists in df
}
\description{
Returns the best longitude column name in df data.frame, if one exists.
}
\keyword{internal}
