% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskedRasterHelper.R
\name{getMaskBand}
\alias{getMaskBand}
\title{Get mask band for masked raster.}
\usage{
getMaskBand(raster)
}
\arguments{
\item{raster}{SpatRaster object containing a single layer}
}
\description{
Get mask band for masked raster.
}
