% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listRaster.R
\name{getMinAndMaxYear}
\alias{getMinAndMaxYear}
\title{Gets the minimum and maximum year for the time dimension values of a WMS Layer.}
\usage{
getMinAndMaxYear(time_dimension_values)
}
\arguments{
\item{time_dimension_values}{A string or character vector of date strings that represent the time dimension e.g. can be fetched with wms_layer$getTimeDimension()$values.
Can be in the format of "2000-01-01T00:00:00.000Z/2019-01-01T00:00:00.000Z/PT1S" (in which the min year will be 2000 and the max year 2019) or several individual dates, 
for which the min and max year will be calculated.}
}
\value{
A named list with a value for min and max.
}
\description{
Gets the minimum and maximum year for the time dimension values of a WMS Layer.
}
\keyword{internal}
