% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPR.R
\name{getPR}
\alias{getPR}
\title{Download PR points from the MAP database}
\usage{
getPR(country = NULL, ISO = NULL, species)
}
\arguments{
\item{country}{string containing name of desired country, e.g. \code{ c("Country1", "Country2", ...)} OR \code{ = "ALL"} (use either \code{country} OR \code{ISO}, not both)}

\item{ISO}{string containing ISO3 code for desired country, e.g. \code{c("XXX", "YYY", ...)} OR \code{ = "ALL"} (use either \code{country} OR \code{ISO}, not both)}

\item{species}{string specifying the Plasmodium species for which to find PR points, options include: \code{"Pf"} OR \code{"Pv"} OR \code{"BOTH"}}
}
\value{
\code{getPR} returns a dataframe containing the below columns, in which each row represents a distinct data point/ study site.

\enumerate{
\item \code{COLUMNNAME} description of contents
\item \code{COLUMNNAME} description of contents
\item \code{COLUMNNAME} description of contents
}
}
\description{
\code{getPR} downloads all publicly available PR points for a specified country (or countries) and returns this as a dataframe.
}
\details{
\code{country} and \code{ISO} refer to countries and a lower-level administrative regions such as Mayotte and Grench Guiana.
}
\examples{
#Download PfPR data for Nigeria and Cameroon and map the locations of these points using autoplot
NGA_CMR_PR <- getPR(country = c("Nigeria", "Cameroon"), species = "Pf")
\dontrun{autoplot(NGA_CMR_PR)}

#Download PfPR data for Madagascar and map the locations of these points using autoplot
Madagascar_pr <- getPR(ISO = "MDG", species = "Pv")
\dontrun{autoplot(Madagascar_pr)}

\dontrun{getPR(country = "ALL", species = "BOTH")}


}
\seealso{
\code{autoplot} method for quick mapping of PR point locations (\code{\link{autoplot.pr.points}}).
}
