% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listShp.R
\name{getPropertiesForAdminLevel}
\alias{getPropertiesForAdminLevel}
\title{Gets the property string to provide to getFeatures input propertyName, given an admin level.}
\usage{
getPropertiesForAdminLevel(admin_level)
}
\arguments{
\item{admin_level}{The admin level as a character - either 'admin0', 'admin1', 'admin2' or 'admin3'}
}
\value{
The properrty string to pass to getFeartures as propertyName
}
\description{
Gets the property string to provide to getFeatures input propertyName, given an admin level.
}
\keyword{internal}
