% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{getRasterDatasetIdFromSurface}
\alias{getRasterDatasetIdFromSurface}
\title{Gets the rasters dataset id, given a surface (title). If more than one rasters have that surface/title, then the one with
the most recent version is selected. If there are no matches, the program will stop with a relevant message.}
\usage{
getRasterDatasetIdFromSurface(rasterList, surface)
}
\arguments{
\item{rasterList}{data.frame containing available raster information, as returned by listRaster().}

\item{surface}{character that is the surface/title of the raster.}
}
\value{
character that is the dataset id of the raster that matches the given surface.
}
\description{
Gets the rasters dataset id, given a surface (title). If more than one rasters have that surface/title, then the one with
the most recent version is selected. If there are no matches, the program will stop with a relevant message.
}
\keyword{internal}
