% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpBbox.R
\name{getSpBbox}
\alias{getSpBbox}
\title{Return sp style bbox}
\usage{
getSpBbox(sfBboxOrShp)
}
\arguments{
\item{sfBboxOrShp}{sf shapefile or result of sf::st_bbox(sf_shp)}
}
\value{
bbox in sp style. A 2x2 matrix - the first column has the minimum, the second the maximum values; rows 1 & 2 represent the x & y dimensions respectively (matrix(c("xmin", "ymin","xmax", "ymax"), nrow = 2, ncol = 2, dimnames = list(c("x", "y"), c("min", "max"))))
}
\description{
Return sp style bbox
}
