% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isAvailable.R
\name{isAvailable}
\alias{isAvailable}
\title{Check whether PR points are available for a given country}
\usage{
isAvailable(country = NULL, ISO = NULL, full_results = FALSE)
}
\arguments{
\item{country}{string containing name of desired country, e.g. \code{ c("Country1", "Country2", ...)} OR \code{ = "ALL"} (use either \code{country} OR \code{ISO}, not both)}

\item{ISO}{string containing ISO3 code for desired country, e.g. \code{c("XXX", "YYY", ...)} OR \code{ = "ALL"} (use either \code{country} OR \code{ISO}, not both)}

\item{full_results}{By default this is FALSE meaning the function only gives a message outlining whether specified country is available, if \code{full_results == TRUE}, the function returns a data.frame outlining data availability.}
}
\value{
\code{isAvailable} returns a list of countries for which PR points are publicly available.

if \code{full_results == TRUE}, a dataframe is returned with the following columns:
\enumerate{
\item \code{available} - those of the specified countries for which PR data points are available
\item \code{not_available}- those of the specified countries for which PR data points are not available
\item \code{possible_match}- agrep-matched country names indicating potential mispellings of countries marked as'not available'
}
}
\description{
\code{isAvailable} checks whether the MAP database contains PR points for the specified country/countries.
}
\examples{
isAvailable(country = "Suriname")
x <- isAvailable(ISO = "NGA", full_results = TRUE)
}
