% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isAvailable_pr.R
\name{isAvailable_pr}
\alias{isAvailable_pr}
\title{Check whether PR points are available for a given location}
\usage{
isAvailable_pr(
  country = NULL,
  ISO = NULL,
  continent = NULL,
  full_results = FALSE,
  version = NULL,
  sourcedata = NULL
)
}
\arguments{
\item{country}{string containing name of desired country, e.g. \code{ c("Country1", "Country2", ...)} (use one of \code{country} OR \code{ISO} OR \code{continent}, not combined)}

\item{ISO}{string containing ISO3 code for desired country, e.g. \code{c("XXX", "YYY", ...)} (use one of \code{country} OR \code{ISO} OR \code{continent}, not combined)}

\item{continent}{string containing name of continent for desired data, e.g. \code{c("Continent1", "Continent2", ...)}(use one of \code{country} OR \code{ISO} OR \code{continent}, not combined)}

\item{full_results}{By default this is FALSE meaning the function only gives a message outlining whether specified country is available, if \code{full_results == TRUE}, the function returns a named list outlining data availability.}

\item{version}{(optional) The PR dataset version to use. If not provided, will just use the most recent version of PR data. (To see available version options, 
use listPRPointVersions)}

\item{sourcedata}{deprecated argument. Please remove it from your code.}
}
\value{
\code{isAvailable_pr} returns a named list of input locations with information regarding data availability.

if \code{full_results == TRUE}, a named list is returned with the following elements:
\enumerate{
\item \code{location} - specified input locations
\item \code{is_available}- 1 or 0; indicating whether data is available for this location
\item \code{possible_match}- agrep-matched country names indicating potential mispellings of countries where is_available == 0; NA if data is available for this location.
}
}
\description{
\code{isAvailable_pr} checks whether the MAP database contains PR points for the specified country/location.
}
\examples{
\dontrun{
isAvailable_pr(country = "Suriname")
x <- isAvailable_pr(ISO = "NGA", full_results = TRUE)
x <- isAvailable_pr(continent = "Oceania", full_results = TRUE)
}
}
