% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskedRasterHelper.R
\name{isMaskedRaster}
\alias{isMaskedRaster}
\title{Returns true if second band of raster is a mask}
\usage{
isMaskedRaster(raster)
}
\arguments{
\item{raster}{SpatRaster object containing a single layer}
}
\description{
Returns true if second band of raster is a mask
}
