% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAdministrativeBoundariesDatasets.R
\name{listAdministrativeBoundariesDatasets}
\alias{listAdministrativeBoundariesDatasets}
\title{List all datasets from the Web Feature Services provided by the Malaria Atlas Project within the Admin Units workspace.}
\usage{
listAdministrativeBoundariesDatasets()
}
\value{
A data.frame with columns 'dataset_id', 'version', and 'workspace' representing the unique identifier, version, and domain/workspace of the datasets.
The version can then be provided to other functions to fetch the data within that dataset. e.g. in getShp
}
\description{
\code{listAdministrativeBoundariesDatasets} lists minimal information of all the feature datasets in the Admin Units workspace 
from the Web Feature Services provided by the Malaria Atlas Project.
}
