% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAll.R
\name{listAll}
\alias{listAll}
\title{List countries with available PfPR data.}
\usage{
listAll()
}
\value{
\code{listAll} returns a list of countries for which PfPR points are publicly available.
}
\description{
\code{listAll} lists all countries for which there are publicly available PfPR datapoints.
}
\examples{
listAll()
}
