% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAllShp.R
\name{listAllShp}
\alias{listAllShp}
\title{List administrative units for which shapefiles are stored on the MAP geoserver.}
\usage{
listAllShp()
}
\value{
\code{listAllShp} returns a data.frame detailing the administrative units for which shapefiles are stored on the MAP geoserver.
}
\description{
\code{listAllShp} lists all administrative units for which shapefiles are stored on the MAP geoserver.
}
\examples{
available_admin_units <- listAllShp()
}
