% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listData.R
\name{listData}
\alias{listData}
\title{List administrative units for which shapefiles are stored on the MAP geoserver.}
\usage{
listData(datatype = NULL, printed = TRUE)
}
\value{
\code{listShp} returns a data.frame detailing the administrative units for which shapefiles are stored on the MAP geoserver.
}
\description{
\code{listShp} lists all administrative units for which shapefiles are stored on the MAP geoserver.
}
\examples{
available_admin_units <- listShp()
}
