% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4r_utils.R
\name{listFeatureTypeDatasetsFromWorkspace}
\alias{listFeatureTypeDatasetsFromWorkspace}
\title{List all datasets from the Web Feature Services provided by the Malaria Atlas Project within the given workspace.}
\usage{
listFeatureTypeDatasetsFromWorkspace(workspace)
}
\arguments{
\item{workspace}{Character vector representing the name of the workspace.}
}
\value{
A data.frame with columns 'id', 'version', and 'workspace' representing the unique identifier, version, and domain of the datasets
}
\description{
This function lists minimal information of all the feature datasets from the Web Feature Services provided by the Malaria Atlas Project, in the given workspace.
}
\keyword{internal}
