% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listParasiteRateDatasets.R
\name{listParasiteRateDatasets}
\alias{listParasiteRateDatasets}
\title{List all datasets from the Web Feature Services provided by the Malaria Atlas Project within the Parasite Rate workspace.}
\usage{
listParasiteRateDatasets()
}
\value{
A data.frame with columns 'dataset_id', 'version', and 'workspace' representing the unique identifier, version, and domain/workspace of the datasets.
The dataset_id can then be provided to other functions to fetch the data within that dataset. e.g. in getPR
}
\description{
\code{listParasiteRateDatasets} lists minimal information of all the feature datasets in the Parasite Rate workspace 
from the Web Feature Services provided by the Malaria Atlas Project.
}
