% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPoints.R
\name{listPoints}
\alias{listPoints}
\title{Deprecated function. Please instead use listPRPointCountries for pr points, and listVecOccPointCountries for vector points}
\usage{
listPoints(printed = TRUE, sourcedata, version = NULL)
}
\description{
\code{listPoints} deprecated function Please remove it from your code.
}
