% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPRPointCountries.R
\name{listPRPointCountries}
\alias{listPRPointCountries}
\title{List countries where there is pr point data available}
\usage{
listPRPointCountries(printed = TRUE, version = NULL)
}
\arguments{
\item{printed}{Should the list be printed to the console?}

\item{version}{(optional) The PR dataset version to use If not provided, will just use the most recent version of PR data. (To see available version options, 
use listPRPointVersions)}
}
\value{
\code{listPRPointCountries} returns a data.frame detailing the countries for which PR points are publicly available.
}
\description{
\code{listPRPointCountries}
}
