% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPRPointVersions.R
\name{listPRPointVersions}
\alias{listPRPointVersions}
\title{List all dataset versions from the Web Feature Services provided by the Malaria Atlas Project within the Parasite Rate workspace.}
\usage{
listPRPointVersions(printed = TRUE)
}
\arguments{
\item{printed}{Should the list be printed to the console?}
}
\value{
A data.frame with column 'version'
The version can then be provided to other functions to fetch the data within that dataset. e.g. in getPR
}
\description{
\code{listPRPointVersions} lists available versions of parasite rate point data from the Web Feature Services provided by the Malaria Atlas Project.
}
\examples{
\dontrun{
prDatasets <- listPRPointVersions()
}
}
