% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listShpVersions.R
\name{listShpVersions}
\alias{listShpVersions}
\title{List all versions of admin unit shapes from the Web Feature Services provided by the Malaria Atlas Project within the Admin Units workspace.}
\usage{
listShpVersions(printed = TRUE)
}
\arguments{
\item{printed}{Should the list be printed to the console?}
}
\value{
A data.frame with column 'version'.
The version can then be provided to other functions to fetch the data within that dataset. e.g. in getShp
}
\description{
\code{listShpVersions} lists available versions of Admin Unit shapefiles from the Web Feature Services provided by the Malaria Atlas Project.
}
\examples{
\dontrun{
vecOccDatasets <- listShpVersions()
}
}
