% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSpecies.R
\name{listSpecies}
\alias{listSpecies}
\title{list all species which have occurrence data within the MAP database.}
\usage{
listSpecies(printed = TRUE)
}
\arguments{
\item{printed}{should the list be printed to the database.}
}
\value{
\code{listSpecies} returns a data.frame detailing the following information for each species available to download from the Malaria Atlas Project database.

\enumerate{
\item \code{species} string detailing species
}
}
\description{
\code{listSpecies} lists all species occurrence data available to download from the Malaria Atlas Project database.
}
\examples{
\donttest{
available_species <- listSpecies()
}

}
