% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVecOccPointCountries.R
\name{listVecOccPointCountries}
\alias{listVecOccPointCountries}
\title{List countries where there is vector occurrence point data available}
\usage{
listVecOccPointCountries(printed = TRUE, version = NULL)
}
\arguments{
\item{printed}{Should the list be printed to the console?}

\item{version}{(optional) The vector occurrence dataset version to use If not provided, will just use the most recent version of vector occurrence data. (To see available version options, 
use listVecOccPointVersions)}
}
\value{
\code{listVecOccPointCountries} returns a data.frame detailing the countries for which vector occurrence points are publicly available.
}
\description{
\code{listVecOccPointCountries}
}
