% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVectorOccurrenceDatasets.R
\name{listVectorOccurrenceDatasets}
\alias{listVectorOccurrenceDatasets}
\title{List all datasets from the Web Feature Services provided by the Malaria Atlas Project within the Vector Occurrence workspace.}
\usage{
listVectorOccurrenceDatasets()
}
\value{
A data.frame with columns 'dataset_id', 'version', and 'workspace' representing the unique identifier, version, and domain/workspace of the datasets.
The dataset_id can then be provided to other functions to fetch the data within that dataset. e.g. in getVecOcc
}
\description{
\code{listVectorOccurrenceDatasets} lists minimal information of all the feature datasets in the Vector Occurrence workspace 
from the Web Feature Services provided by the Malaria Atlas Project.
}
