% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplotRasterHelper.R
\name{makeSpatRasterAutoplot}
\alias{makeSpatRasterAutoplot}
\title{Create a single (sub) plot for a SpatRaster}
\usage{
makeSpatRasterAutoplot(
  spatraster,
  rastername,
  shp_df,
  legend_title,
  fill_scale_transform,
  fill_colour_palette,
  plot_titles
)
}
\arguments{
\item{spatraster}{SpatRaster object containing a single layer}

\item{rastername}{raster name, to include in title}

\item{shp_df}{sf shapefile}

\item{legend_title}{title for legend}

\item{fill_scale_transform}{scale}

\item{fill_colour_palette}{palette}

\item{plot_titles}{bool, whether to include title}
}
\value{
ggplot object
}
\description{
Create a single (sub) plot for a SpatRaster
}
