% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPdata.R
\docType{package}
\name{MAPdata}
\alias{MAPdata}
\alias{MAPdata-package}
\title{Download publicly available Parasite Rate Data from the Malaria Atlas Project geoserver.}
\description{
\code{MAPdata} provides a suite of tools to allow you to
 download all publicly available PfPR points for a specified country
 (or ALL countries) into R as a dataframe.
}
\section{MAPdata functions}{

 \enumerate {
 \item listAll - what does it do?
 \item getPR - what does it do?}
}

