% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.MAPraster.R
\name{autoplot.MAPraster}
\alias{autoplot.MAPraster}
\title{Quickly visualise Rasters downloaded from MAP}
\usage{
\method{autoplot}{MAPraster}(object, shp_df = NULL, legend_title = "",
  plot_title = "", log_scale = FALSE, printed = TRUE)
}
\arguments{
\item{object}{MAPraster object to be visualised.}

\item{shp_df}{Shapefile(s) (data.frame) to plot with downloaded raster.}

\item{legend_title}{String used as title for all colour scale legends.}

\item{log_scale}{Logical vector indicating whether to use a log scale for legend colour scales.}

\item{printed}{Logical vector indicating whether to print maps of supplied rasters.}

\item{page_title}{String used as header for all plot pages.}
}
\value{
\code{autoplot.MAPraster} returns a list of plots (gg objects) for each supplied raster.
}
\description{
\code{autoplot.MAPraster} creates a map of all rasters in a MAPraster object and displays these in a grid.
}
\examples{
#Download PfPR2-10 Raster for Madagascar in 2016 and visualise this on a map.
\dontrun{MDG_shp <- getShp(ISO = "MDG", admin_level = "admin0")
MDG_PfPR2_10 <- getRaster(surface = "PfPR2-10", shp = MDG_shp, year = c(2010:2015))
MDG_PfPR2_10 <- as.MAPraster(MDG_PfPR2_10)
autoplot(MDG_PfPR2_10)}

#Download global raster of G6PD deficiency from Howes et al 2012 and visualise this on a map.
\dontrun{G6PDd_global <- getRaster(surface = "G6PD Deficiency")
G6PDd_global <- as.MAPraster(G6PDd_global)
autoplot(G6PDd_global)}

}
\seealso{
\code{\link{getRaster}}:

to download rasters directly from MAP.

\code{\link{as.MAPraster}}:

to convert RasterLayer/RasterStack objects into a 'MAPraster' object (data.frame) for easy plotting with ggplot.

\code{\link{autoplot.MAPraster}}:

to quickly visualise MAPraster objects created using /code{as.MAPraster}.
}
