% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.MAPshp.R
\name{autoplot.MAPshp}
\alias{autoplot.MAPshp}
\title{Create a basic plot to visualise downloaded shapefiles}
\usage{
\method{autoplot}{MAPshp}(object, map_title = NULL, facet = FALSE)
}
\arguments{
\item{object}{a MAPShp object downloaded using /code{/link{getShp}} with format = "df" specified.}

\item{map_title}{custom title used for the plot}

\item{facet}{if TRUE, splits map into a separate facet for each malaria species}
}
\value{
\code{autoplot.MAPshp} returns a map (gg object) of the supplied MAPShp dataframe.
}
\description{
\code{autoplot.MAPshp} creates a map of shapefiles downloaded using getShp.
}
\examples{
\dontrun{MDG_shp <- getShp(ISO = "MDG", admin_level = "admin0")}
\dontrun{autoplot(MDG_shp)}

}
