% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRplot.R
\name{autoplot.pr.points}
\alias{autoplot.pr.points}
\title{Create a basic plot showing locations of downloaded PR points}
\usage{
autoplot.pr.points(pr_data, col_both = "orchid3", col_confidential = "grey",
  col_pf = "royalblue3", col_pv = "coral", map_title = NULL, ...)
}
\description{
\code{text} more text
}
\examples{
\dontrun{pr_data <- getPR(country = c("Nigeria"), species = "Pf")
autoplot(pr_data)}
}
