% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.pr.points.R
\name{autoplot.pr.points}
\alias{autoplot.pr.points}
\title{Create a basic plot showing locations of downloaded PR points}
\usage{
\method{autoplot}{pr.points}(object, col_both = "orchid3",
  col_confidential = "grey", col_pf = "royalblue3", col_pv = "coral",
  map_title = NULL, extent = "national_only", facet = FALSE,
  hide_confidential = FALSE, ...)
}
\description{
\code{text} more text
}
\examples{
\dontrun{object <- getPR(country = c("Nigeria"), species = "Pf")}
\dontrun{autoplot(object)}
}
