% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_MAPraster.R
\name{autoplot_MAPraster}
\alias{autoplot_MAPraster}
\title{Quickly visualise Rasters downloaded from MAP}
\usage{
autoplot_MAPraster(object, ...)
}
\arguments{
\item{object}{RasterLayer/RasterStack object to be visualised.}

\item{...}{other optional arguments to autoplot.MAPraster (e.g. shp_df, legend_title, page_title...)}
}
\value{
\code{autoplot_MAPraster} returns a list of plots (gg objects) for each supplied raster.
}
\description{
\code{autoplot_MAPraster} is a wrapper for /code{/link{autoplot.MAPraster}} that calls /code{/link{as.MAPraster}} to allow automatic map creation for RasterLayer/RasterStack objects downloaded from MAP.
}
\examples{
#Download PfPR2-10 Raster for Madagascar in 2016 and visualise this on a map.
\dontrun{MDG_shp <- getShp(ISO = "MDG", admin_level = "admin0")
MDG_PfPR2_10 <- getRaster(surface = "PfPR2-10", shp = MDG_shp, year = 2016)
autoplot_MAPraster(MDG_PfPR2_10)}

#Download global raster of G6PD deficiency from Howes et al 2012 and visualise this on a map.
\dontrun{G6PDd_global <- getRaster(surface = "G6PD Deficiency")
autoplot_MAPraster(G6PDd_global)}

}
\seealso{
\code{\link{getRaster}}:

to download rasters directly from MAP.

\code{\link{as.MAPraster}}:

to convert RasterLayer/RasterStack objects into a 'MAPraster' object (data.frame) for easy plotting with ggplot.

\code{\link{autoplot.MAPraster}}:

to quickly visualise MAPraster objects created using /code{as.MAPraster}.
}
