% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPfPR.R
\name{getPfPR}
\alias{getPfPR}
\title{Download publicly available PfPR points from MAP's geoserver.}
\usage{
getPfPR(country)
}
\arguments{
\item{country}{\code{ = c("Country1", "Country2", ...)} OR \code{ = "ALL"}

specifies which country/countries you wish to include in your download}
}
\value{
\code{getPfPR} returns a dataframe containing the below columns, in which each row represents a distinct data point/ study site.

\enumerate{
\item \code{COLUMNNAME} description of contents
\item \code{COLUMNNAME} description of contents
\item \code{COLUMNNAME} description of contents
}
}
\description{
\code{getPfPR} downloads all publicly available PfPR points for a specified country and returns this as a dataframe.
}
\examples{
getPfPR(country = c("Nigeria", "Kenya"))
getPfPR(country = "ALL")
}
