% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPR.R
\name{getPR}
\alias{getPR}
\title{Download publicly available PR points from MAP's geoserver.}
\usage{
getPR(country = NULL, ISO = NULL, species)
}
\arguments{
\item{country}{\code{ = c("Country1", "Country2", ...)} OR \code{ = "ALL"}

specifies which country/countries you wish to include in your download}
}
\value{
\code{getPR} returns a dataframe containing the below columns, in which each row represents a distinct data point/ study site.

\enumerate{
\item \code{COLUMNNAME} description of contents
\item \code{COLUMNNAME} description of contents
\item \code{COLUMNNAME} description of contents
}
}
\description{
\code{getPR} downloads all publicly available PR points for a specified country and returns this as a dataframe.
}
\examples{
getPR(country = c("Nigeria", "Kenya"), species = "Pf")
getPR(country = "ALL", species = "Pv")
}
