% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRaster.R
\name{getRaster}
\alias{getRaster}
\title{Documentation}
\usage{
getRaster(surface = "PfPR2-10", shp = NULL, view_bbox = NULL,
  file_path = tempdir(), format = "raster", year = NULL)
}
\description{
Documentation
}
