% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRaster.R
\name{getRaster}
\alias{getRaster}
\title{Download Rasters produced by the Malaria Atlas Project}
\usage{
getRaster(surface = "PfPR2-10", shp = NULL, view_bbox = NULL,
  file_path = tempdir(), year = rep(NA, length(surface)))
}
\arguments{
\item{surface}{string containing code for desired raster(s), e.g. \code{c("rasterCode1", "rasterCode2")}. Defaults to "PfPR2-10" - the most recent global raster of PfPR 2-10.}

\item{shp}{SpatialPolygon(s) object of a shapefile to use when clipping downloaded rasters. (use either \code{shp} OR \code{view_bbox}; if neither is specified global raster is returned).}

\item{view_bbox}{matrix containing bounding box for which rasters will be downloaded (as returned by sp::bbox()) (use either \code{shp} OR \code{view_bbox}; if neither is specified global raster is returned).}

\item{file_path}{string specifying the directory to which working files will be downloaded. Defaults to tempdir().}

\item{year}{default = \code{rep(NA, length(surface))} (use \code{NA} for static rasters); for time-varying rasters: if downloading a single surface for one or more years, \code{year} should be a string specifying the desired year(s). if downloading more than one surface, use a list the same length as \code{surface}, providing the desired year-range for each time-varying surface in \code{surface} or \code{NA} for static rasters.}
}
\value{
\code{getRaster} returns a RasterLayer (if only a single raster is queried) or RasterStack (for multiple rasters) for the specified extent.
}
\description{
\code{getRaster} downloads publicly available MAP rasters for a specific surface & year, clipped to a provided bounding box or shapefile.
}
\examples{
#Download PfPR2-10 Raster for Madagascar in 2016 and visualise this immediately.
\dontrun{
MDG_shp <- getShp(ISO = "MDG", admin_level = "admin0")
MDG_PfPR2_10 <- getRaster(surface = "PfPR2-10", shp = MDG_shp, year = 2016)
autoplot_MAPraster(MDG_PfPR2_10)}

#Download global raster of G6PD deficiency from Howes et al 2012.
\dontrun{G6PDd_global <- getRaster(surface = "G6PD Deficiency")
autoplot_MAPraster(G6PDd_global)}

}
\seealso{
(\code{\link{autoplot_MAPraster}}:

to quickly visualise rasters downloded using /code{getRaster}.

\code{\link{as.MAPraster}}:

to convert RasterLayer/RasterStack objects into a 'MAPraster' object (data.frame) for easy plotting with ggplot.

\code{\link{autoplot.MAPraster}}:

to quickly visualise MAPraster objects created using /code{as.MAPraster}.
}
