% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShp.R
\name{getShp}
\alias{getShp}
\title{Download MAPadmin2013 Administrative Boundary Shapefiles from the MAP geoserver}
\usage{
getShp(object = NULL, country = NULL, ISO = NULL, admin_level = "both",
  extent = "national_only")
}
\arguments{
\item{object}{object of class pr.points (e.g. data downloaded uisng getPR()) for which corresponding shapefiles are desired (use either \code{object} OR \code{ISO} OR \code{country}, not in combination)}

\item{country}{string containing name of desired country, e.g. \code{ c("Country1", "Country2", ...)} OR \code{ = "ALL"} (use either \code{object} OR \code{ISO} OR \code{country}, not in combination)}

\item{ISO}{string containing ISO3 code for desired country, e.g. \code{c("XXX", "YYY", ...)} OR \code{ = "ALL"} (use either \code{object} OR \code{ISO} OR \code{country}, not in combination)}

\item{admin_level}{string specifying the administrative level for which shapefile are desired (only "admin1", "admin0", or "both" accepted)}

\item{extent}{string specifying method of defining shapefile extent - accepts either \code{bbox} (for use with pr.points object only - defines 2% bounding box around data and downloads shapefiles intersecting this box) or "national_only" (the default - downloads all shapefiles associated with specified country name or names of countries within 'object')}
}
\value{
\code{getShp} returns a list containing separate shapefile dataframes for each administrative level. The following attribute fields are included:

\enumerate{
\item \code{COLUMNNAME} description of contents
\item \code{COLUMNNAME} description of contents
\item \code{COLUMNNAME} description of contents
}
}
\description{
\code{getShp} downloads shapefiles for a specified country (or countries) and returns this as a spatial
}
\examples{
#Download PfPR data & associated shapefiles for Nigeria and Cameroon
\dontrun{NGA_CMR_PR <- getPR(country = c("Nigeria", "Cameroon"), species = "Pf")}
\dontrun{NGA_CMR_shp <- getShp(country = c("Nigeria", "Cameroon"))}

#' #Download PfPR data & associated shapefiles for Chad
\dontrun{Chad_PR <- getPR(ISO = "TCD", species = "both")}
\dontrun{Chad_shp <- getShp(ISO = "TCD")}

#' #Download PfPR data & associated shapefiles defined by bbox for Madagascar
\dontrun{MDG_PR <- getPR(country = "Madagascar", species = "Pv")}
\dontrun{MDG_shp <- getShp(object = MDG_PR, extent = "bbox")}


}
\seealso{
\code{autoplot} method for quick mapping of PR point locations (\code{\link{autoplot.pr.points}}).
}
