% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableData.R, R/isAvailable.R
\name{isAvailable}
\alias{isAvailable}
\title{Available data to download from the MAP geoserver.}
\usage{
isAvailable(sourcedata = NULL, full_results = FALSE, country = NULL,
  ISO = NULL, continent = NULL)

isAvailable(sourcedata = NULL, full_results = FALSE, country = NULL,
  ISO = NULL, continent = NULL)
}
\arguments{
\item{sourcedata}{One of 'pr points' or 'vector points'}

\item{full_results}{Should the list be printed to the console?}

\item{sourcedata}{One of 'pr points' or 'vector points'}

\item{full_results}{Should the list be printed to the console?}

\item{...}{Other arguments to be passed to \code{isAvailable*} functions. (e.g. \code{continent})}
}
\value{
\code{isAvailable} returns a data.frame detailing the administrative units for which shapefiles are stored on the MAP geoserver.

\code{isAvailable} returns a data.frame detailing the administrative units for which shapefiles are stored on the MAP geoserver.
}
\description{
\code{isAvaiable} is a wrapper for isAvailable_pr and isAvailable_vec, listing data (PR survey point location data and vector occurrence locations available to download from the MAP geoserver.

\code{isAvaiable} is a wrapper for isAvailable_pr and isAvailable_vec, listing data (PR survey point location data and vector occurrence locations available to download from the MAP geoserver.
}
\examples{
\donttest{
available_pr_locations <- isAvailable_pr()
available_vector_locations <- isAvailable_vec()
}
\donttest{
available_pr_locations <- isAvailable_pr()
available_vector_locations <- isAvailable_vec()
}
}
\seealso{
\code{link{isAvailable_pr}}
\code{\link{isAvailable_vec}}

\code{link{isAvailable_pr}}
\code{\link{isAvailable_vec}}
}
