% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAll.R
\name{listAll}
\alias{listAll}
\title{List countries with available PR data.}
\usage{
listAll(printed = TRUE)
}
\value{
\code{listAll} returns a list of countries for which PR points are publicly available.
}
\description{
\code{listAll} lists all countries for which there are publicly visible PR datapoints in the MAP database.
}
\examples{
listAll()
}
