% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listData.R
\name{listData}
\alias{listData}
\title{Deprecated function. Please instead use listPRPointCountries for pr points, listVecOccPointCountries for vector points, listRaster for raster and listShp for shape.}
\usage{
listData(datatype, printed = TRUE, ...)
}
\arguments{
\item{datatype}{"pr points", "vector points" "raster", or "shape"}

\item{printed}{whether to pretty print the output in console}

\item{...}{passed on to listPRPointCountries, listVecOccPointCountries, listShp}
}
\description{
\code{listData}  deprecated function Please remove it from your code.
}
