% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPoints.R
\name{listPoints}
\alias{listPoints}
\title{List countries with available MAP Point data.}
\usage{
listPoints(printed = TRUE, sourcedata, dataset_id = NULL)
}
\arguments{
\item{printed}{Should the list be printed to the console?}

\item{sourcedata}{String contining desired dataset within the Malaria Atlas database to be searched, e.g \code{"pr points"} OR \code{"vector points"}}

\item{dataset_id}{A character string specifying the dataset ID. Is NULL by default, and the most recent version of the pr points or vector points will be selected.}
}
\value{
\code{listPoints} returns a data.frame detailing the countries for which PR points are publicly available.
}
\description{
\code{listPoints} lists all countries for which there are publicly visible datapoints in the MAP database required.
}
\examples{
\donttest{
listPoints(sourcedata = "pr points")
listPoints(sourcedata = "vector points")
}
}
