% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPoints.R
\name{listPoints}
\alias{listPoints}
\title{List countries with available PR Point data.}
\usage{
listPoints(printed = TRUE)
}
\value{
\code{listPoints} returns a data.frame detailing the countries for which PR points are publicly available.
}
\description{
\code{listPoints} lists all countries for which there are publicly visible PR datapoints in the MAP database.
}
\examples{
listPoints()
}
