% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPoints.R
\name{listPoints}
\alias{listPoints}
\title{Deprecated function. Please instead use listPRPointCountries for pr points, and listVecOccPointCountries for vector points}
\usage{
listPoints(printed = TRUE, sourcedata, version = NULL)
}
\arguments{
\item{printed}{whether to pretty print the output in console}

\item{sourcedata}{"pr points" or "vector points"}

\item{version}{(optional) The PR dataset version to use If not provided, will just use the most recent version of PR data. (To see available version options, 
use listPRPointVersions)}
}
\description{
\code{listPoints} deprecated function Please remove it from your code.
}
