% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPPfPR.R
\docType{package}
\name{MAPPFPR}
\alias{MAPPFPR}
\alias{MAPPFPR-package}
\title{Download publicly available Parasite Rate Data from the Malaria Atlas Project geoserver.}
\description{
\code{MAPPR} provides a suite of tools to allow you to
 download all publicly available PfPR points for a specified country
 (or ALL countries) into R as a dataframe.
}
\section{MAPPfPR functions}{

 \enumerate {
 \item listAll - what does it do?
 \item getPfPR - what does it do?}
}

